Neia V1.09: Quality of Life/Preperation update

+++ commands/delete.js
@@ -1,24 +1,28 @@
 module.exports = {
 	name: 'delete',
-	description: 'Delete messages in bulk.',
-	admin: true,
+	summary: 'Delete messages in bulk',
+	description: 'Delete up too 100 messages in bulk.',
+	category: 'admin',
 	aliases: ['remove'],
 	args: false,
-	usage: '(amount of messaged to delete)',
-	owner: false,
+	usage: '<amount>',
 	cooldown: 4,
-	music: false,
 
-	execute(msg, args, profile, bot, options, ytAPI, logger, cooldowns) {
+	execute(msg, args, profile, guildProfile, bot, options, ytAPI, logger, cooldowns) {
+
+		const amount = args[0];
+		if (isNaN(amount)) return msg.channel.send(`**${amount}** is not a valid number`);
+		if (amount < 1 || amount > 100) return msg.channel.send('Input a number between 1 and 100');
+
 
 		try {
 			msg.delete();
-			msg.channel.bulkDelete(args[0]);
-			logger.log('info', `${msg.author.tag} deleted ${args[0]} messages in channel ${msg.channel.name}`);
-		} 
-		catch (e) {
+			msg.channel.bulkDelete(amount);
+			logger.log('info', `*${msg.author.tag}* deleted **${amount}** messages in channel ${msg.channel.name}`);
+		}
+		catch (error) {
 			msg.channel.send('Something went wrong');
-			logger.log('error', e);
+			return logger.error(error.stack);
 		}
 	},
 };
